requirements.txt
Flask==2.1.1
pandas==1.4.2
spotipy==2.19.0
PyYAML==5.4.1
scikit-learn==1.0.2

